heroku デプロイ用

+++ src/main/kotlin/kissybnts/fbmmessengerbot/model/FBMessengerBotWebhookEntry.kt
@@ -5,5 +5,5 @@ import java.math.BigDecimal
 /**
  * Created by kishida on 2016/04/17.
  */
-open class FBMessengerBotWebhookEntry(val id: BigDecimal = BigDecimal(0), val time: BigDecimal = BigDecimal(0), val messaging: MutableList<FBMessengerBotWebhookEntryMessaging> = mutableListOf()) {
+open class FBMessengerBotWebhookEntry(val id: Long = 0, val time: Long = 0, val messaging: MutableList<FBMessengerBotWebhookEntryMessaging> = mutableListOf()) {
 }
