+++ src/main/kotlin/kissybnts/fbmmessengerbot/model/FBMessengerBotWebhookEntryMessaging.kt
@@ -5,5 +5,5 @@ import java.math.BigDecimal
 /**
  * .
  */
-data class FBMessengerBotWebhookEntryMessaging(val sender: Map<String, BigDecimal> = mapOf(), val  recipient: Map<String, BigDecimal> = mapOf(), val timestamp: BigDecimal = BigDecimal(0), val message: FBMessengerBotWebhookEntryMessagingMessage = FBMessengerBotWebhookEntryMessagingMessage()) {
+data class FBMessengerBotWebhookEntryMessaging(val sender: Map<String, Long> = mapOf(), val  recipient: Map<String, Long> = mapOf(), val timestamp: Long = 0, val message: FBMessengerBotWebhookEntryMessagingMessage = FBMessengerBotWebhookEntryMessagingMessage()) {
 }
