+++ src/main/kotlin/kissybnts/fbmmessengerbot/model/FBMessengerBotWebhookRecipient.kt
@@ -6,6 +6,6 @@ import java.math.BigDecimal
 /**
  * Created by kishida on 2016/04/17.
  */
-data class FBMessengerBotWebhookRecipient(var recipient: Map<String, BigDecimal> = mapOf(),
+data class FBMessengerBotWebhookRecipient(var recipient: Map<String, Long> = mapOf(),
                                           var message: Map<String, String> = mapOf()) {
 }
